+++ src/config/layers.js
@@ -1,16 +1,12 @@
 import proj4 from 'proj4';
 import { register } from 'ol/proj/proj4';
 import { Layer } from 'mobility-toolbox-js/ol';
-import GeometryType from 'ol/geom/GeometryType';
 import MapboxStyleLayer from '../layers/MapboxStyleLayer';
 import TrafimageMapboxLayer from '../layers/TrafimageMapboxLayer';
 import ZweitausbildungAbroadLayer from '../layers/ZweitausbildungAbroadLayer';
 import ZweitausbildungPoisLayer from '../layers/ZweitausbildungPoisLayer';
 import ZweitausbildungRoutesLayer from '../layers/ZweitausbildungRoutesLayer';
 import ZweitausbildungRoutesHighlightLayer from '../layers/ZweitausbildungRoutesHighlightLayer';
-import BeleuchtungsLayer from '../layers/BeleuchtungsLayer';
-import MapsGeoAdminLayer from '../layers/MapsGeoAdminLayer';
-import LevelLayer from '../layers/LevelLayer';
 import { dataLayer } from './ch.sbb.netzkarte';
 
 proj4.defs(
@@ -28,29 +24,6 @@ proj4.defs(
 );
 
 register(proj4);
-
-export const beleuchtungDataLayer = new TrafimageMapboxLayer({
-  name: 'ch.sbb.beleuchtungsstaerken.data',
-  visible: true,
-  preserveDrawingBuffer: true,
-  isBaseLayer: true,
-  isQueryable: false,
-  zIndex: -1, // Add zIndex as the MapboxLayer would block tiled layers (buslines)
-  style: 'netzkarte_eisenbahninfrastruktur_v3_ch.sbb.beleuchtung',
-  properties: {
-    hideInLegend: true,
-  },
-});
-
-export const netzkarteIsb = new TrafimageMapboxLayer({
-  name: 'ch.sbb.isb',
-  isBaseLayer: true,
-  visible: true,
-  isQueryable: false,
-  preserveDrawingBuffer: true,
-  zIndex: -1,
-  style: 'netzkarte_eisenbahninfrastruktur_v3_ch.sbb.infrastrukturbetreiber',
-});
 
 export const zweitausbildungAbroad = new ZweitausbildungAbroadLayer({
   name: 'ch.sbb.zweitausbildung.abroad',
@@ -333,402 +306,4 @@ export const zweitausbildungRoutes = new Layer({
       ],
     }),
   ],
-});
-
-export const anlagenverantwortliche = new TrafimageMapboxLayer({
-  name: 'ch.sbb.anlagenverantwortliche',
-  style: 'netzkarte_eisenbahninfrastruktur_v3_ch.sbb.anlagenverantwortliche',
-  isBaseLayer: false,
-  visible: true,
-  isQueryable: false,
-  preserveDrawingBuffer: true,
-  zIndex: -1,
-  properties: {
-    hideInLegend: true,
-  },
-});
-
-export const regionenkartePublicSegment = new Layer({
-  name: 'ch.sbb.regionenkarte.intern.av_segmente.public',
-  visible: true,
-  isQueryable: false,
-  properties: {
-    hasInfos: true,
-    layerInfoComponent: 'RegionenkartePublicLayerInfo',
-  },
-  children: [
-    new MapboxStyleLayer({
-      name: 'ch.sbb.regionenkarte.lines',
-      isQueryable: true,
-      mapboxLayer: anlagenverantwortliche,
-      styleLayersFilter: ({ id }) => {
-        return /anlagenverantwortliche\.lines$/.test(id);
-      },
-      featureInfoFilter: (feature) => {
-        // There is some points in this data source and we don't want them.
-        return (
-          feature.getGeometry().getType() === GeometryType.LINE_STRING ||
-          feature.getGeometry().getType() === GeometryType.MULTI_LINE_STRING
-        );
-      },
-      properties: {
-        hideInLegend: true,
-        useOverlay: true, // instead of a Popup , on click an Overlay will be displayed.
-        popupComponent: 'RegionenkarteSegmentPopup',
-      },
-    }),
-    new MapboxStyleLayer({
-      name: 'ch.sbb.regionenkarte.stations',
-      isQueryable: false,
-      mapboxLayer: anlagenverantwortliche,
-      styleLayersFilter: ({ id }) => {
-        return /anlagenverantwortliche\.stations/.test(id);
-      },
-      properties: {
-        hideInLegend: true,
-      },
-    }),
-    new MapboxStyleLayer({
-      name: 'ch.sbb.regionenkarte.regionintersection',
-      isQueryable: true,
-      mapboxLayer: anlagenverantwortliche,
-      styleLayersFilter: ({ id }) => {
-        return /anlagenverantwortliche\.regionintersection/.test(id);
-      },
-      properties: {
-        hideInLegend: true,
-        showPopupOnHover: true,
-        popupComponent: 'RegionenkarteIntersectionPopup',
-      },
-    }),
-  ],
-});
-
-export const regionenkarteOverlayGroup = new Layer({
-  name: 'ch.sbb.infrastruktur.overlay.group',
-  visible: true,
-  isQueryable: false,
-  properties: {
-    hasInfos: true,
-    description: 'ch.sbb.infrastruktur.overlay.group-desc',
-  },
-  children: [
-    new MapboxStyleLayer({
-      name: 'ch.sbb.infrastruktur.betriebspunkte',
-      visible: true,
-      isQueryable: false,
-      mapboxLayer: anlagenverantwortliche,
-      styleLayersFilter: ({ id }) => {
-        // We select all stations
-        return /FanasStation/.test(id);
-      },
-      properties: {
-        hasInfos: true,
-        description: 'ch.sbb.infrastruktur.betriebspunkte-desc',
-      },
-    }),
-    new MapboxStyleLayer({
-      name: 'ch.sbb.infrastruktur.line_point',
-      visible: true,
-      isQueryable: false,
-      mapboxLayer: anlagenverantwortliche,
-      styleLayersFilter: ({ id }) => {
-        return /FanasLine|DFA/.test(id);
-      },
-      properties: {
-        hasInfos: true,
-        description: 'ch.sbb.infrastruktur.line_point-desc',
-      },
-    }),
-  ],
-});
-
-export const netzentwicklungDataLayer = new TrafimageMapboxLayer({
-  name: 'ch.sbb.netzkarte.data',
-  isQueryable: false,
-  preserveDrawingBuffer: true,
-  zIndex: -1,
-  style: 'netzkarte_eisenbahninfrastruktur_v3_ch.sbb.netzentwicklung',
-  properties: {
-    hideInLegend: true,
-  },
-});
-
-export const netzentwicklungProgrammManagerLayer = new MapboxStyleLayer({
-  name: 'ch.sbb.netzentwicklung.programm_manager',
-  mapboxLayer: netzentwicklungDataLayer,
-  visible: false,
-  styleLayersFilter: ({ id }) => /programm_manager$/.test(id),
-  properties: {
-    radioGroup: 'netzentwicklung',
-    popupComponent: 'NetzentwicklungPopup',
-    netzentwicklungRoleType: 'Programm Manager', // display only roles of this type
-    hasInfos: true,
-    useOverlay: true,
-    layerInfoComponent: 'NetzentwicklungLayerInfo',
-  },
-});
-
-export const netzentwicklungSkPlanerLayer = new MapboxStyleLayer({
-  name: 'ch.sbb.netzentwicklung.sk_planer',
-  mapboxLayer: netzentwicklungDataLayer,
-  visible: true,
-  styleLayersFilter: ({ id }) => /sk_planer$/.test(id),
-  properties: {
-    radioGroup: 'netzentwicklung',
-    popupComponent: 'NetzentwicklungPopup',
-    netzentwicklungRoleType: 'S&K Planer', // display only roles of this type
-    hasInfos: true,
-    useOverlay: true,
-    layerInfoComponent: 'NetzentwicklungLayerInfo',
-  },
-});
-
-export const netzentwicklungStrategischLayer = new MapboxStyleLayer({
-  name: 'ch.sbb.netzentwicklung.strategisch',
-  mapboxLayer: netzentwicklungDataLayer,
-  visible: false,
-  styleLayersFilter: ({ id }) => /strategisch$/.test(id),
-  properties: {
-    radioGroup: 'netzentwicklung',
-    popupComponent: 'NetzentwicklungPopup',
-    netzentwicklungRoleType: 'Netzentwickler Strategisch', // display only roles of this type
-    hasInfos: true,
-    useOverlay: true,
-    layerInfoComponent: 'NetzentwicklungLayerInfo',
-  },
-});
-
-export const beleuchtungstaerken1Layer = new BeleuchtungsLayer({
-  name: 'ch.sbb.beleuchtungsstaerken1',
-  mapboxLayer: beleuchtungDataLayer,
-  styleLayersFilter: ({ metadata }) => metadata && metadata.rte_klasse === '1',
-  properties: {
-    hasInfos: true,
-    useOverlay: true,
-    popupComponent: 'BeleuchtungsPopup',
-    layerInfoComponent: 'BeleuchtungLayerInfo',
-  },
-});
-
-export const beleuchtungstaerken2aLayer = new BeleuchtungsLayer({
-  name: 'ch.sbb.beleuchtungsstaerken2a',
-  mapboxLayer: beleuchtungDataLayer,
-  styleLayersFilter: ({ metadata }) => metadata && metadata.rte_klasse === '2a',
-  properties: {
-    hasInfos: true,
-    useOverlay: true,
-    popupComponent: 'BeleuchtungsPopup',
-    layerInfoComponent: 'BeleuchtungLayerInfo',
-  },
-});
-
-export const beleuchtungstaerken2bLayer = new BeleuchtungsLayer({
-  name: 'ch.sbb.beleuchtungsstaerken2b',
-  mapboxLayer: beleuchtungDataLayer,
-  styleLayersFilter: ({ metadata }) => metadata && metadata.rte_klasse === '2b',
-  properties: {
-    hasInfos: true,
-    useOverlay: true,
-    popupComponent: 'BeleuchtungsPopup',
-    layerInfoComponent: 'BeleuchtungLayerInfo',
-  },
-});
-
-export const beleuchtungstaerken3Layer = new BeleuchtungsLayer({
-  name: 'ch.sbb.beleuchtungsstaerken3',
-  mapboxLayer: beleuchtungDataLayer,
-  styleLayersFilter: ({ metadata }) => metadata && metadata.rte_klasse === '3',
-  properties: {
-    hasInfos: true,
-    useOverlay: true,
-    popupComponent: 'BeleuchtungsPopup',
-    layerInfoComponent: 'BeleuchtungLayerInfo',
-  },
-});
-
-export const beleuchtungstaerken4Layer = new BeleuchtungsLayer({
-  name: 'ch.sbb.beleuchtungsstaerken4',
-  mapboxLayer: beleuchtungDataLayer,
-  styleLayersFilter: ({ metadata }) => metadata && metadata.rte_klasse === '4',
-  properties: {
-    hasInfos: true,
-    useOverlay: true,
-    popupComponent: 'BeleuchtungsPopup',
-    layerInfoComponent: 'BeleuchtungLayerInfo',
-  },
-});
-
-const createMapsGeoAdminStyleLayer = (key) => {
-  return new MapboxStyleLayer({
-    name: key,
-    key,
-    visible: false,
-    mapboxLayer: beleuchtungDataLayer,
-    styleLayersFilter: ({ id }) => id === key,
-    properties: {
-      legendKey: key,
-      hasInfos: true,
-      layerInfoComponent: 'MapsGeoAdminLayerInfo',
-    },
-  });
-};
-
-const mapsGeoAdminSchutzgebieteLayerKeys = [
-  'ch.bafu.wrz-jagdbanngebiete_select',
-  'ch.bafu.wrz-wildruhezonen_portal',
-  'ch.bafu.waldreservate',
-  'ch.bafu.unesco-weltnaturerbe',
-  'ch.bak.schutzgebiete-unesco_weltkulturerbe',
-  'ch.bafu.schutzgebiete-smaragd',
-  'ch.bafu.schutzgebiete-paerke_nationaler_bedeutung_perimeter',
-  'ch.bafu.schutzgebiete-ramsar',
-  'ch.pronatura.waldreservate',
-  'ch.pronatura.naturschutzgebiete',
-  'ch.bafu.schutzgebiete-biosphaerenreservate',
-];
-
-export const beleuchtungstaerkenSchutzgebieteLayer = new MapsGeoAdminLayer({
-  name: 'ch.sbb.beleuchtungsstaerken.bafu-schutzgebiete.group',
-  visible: false,
-  children: mapsGeoAdminSchutzgebieteLayerKeys.map(
-    createMapsGeoAdminStyleLayer,
-  ),
-  properties: {
-    featureInfoEventTypes: ['singleclick'],
-    useOverlay: true,
-    popupComponent: 'MapsGeoAdminPopup',
-  },
-});
-
-const mapsGeoAdminBundesinventareLayerKeys = [
-  'ch.bafu.bundesinventare-vogelreservate',
-  'ch.bafu.bundesinventare-auen_vegetation_alpin',
-  'ch.bafu.bundesinventare-trockenwiesen_trockenweiden_anhang2',
-  'ch.bafu.bundesinventare-trockenwiesen_trockenweiden',
-  'ch.bafu.bundesinventare-moorlandschaften',
-  'ch.bafu.bundesinventare-bln',
-  'ch.bafu.bundesinventare-jagdbanngebiete',
-  'ch.bafu.bundesinventare-hochmoore',
-  'ch.bafu.bundesinventare-flachmoore',
-  'ch.bafu.bundesinventare-auen_anhang2',
-  'ch.bafu.bundesinventare-auen',
-  'ch.bafu.bundesinventare-amphibien_wanderobjekte',
-  'ch.bafu.bundesinventare-amphibien',
-  'ch.bafu.bundesinventare-amphibien_anhang4',
-];
-
-export const beleuchtungstaerkenBundesInventareLayer = new MapsGeoAdminLayer({
-  name: 'ch.sbb.beleuchtungsstaerken.bafu-bundesinventare.group',
-  visible: false,
-  children: mapsGeoAdminBundesinventareLayerKeys.map(
-    createMapsGeoAdminStyleLayer,
-  ),
-  properties: {
-    featureInfoEventTypes: ['singleclick'],
-    useOverlay: true,
-    popupComponent: 'MapsGeoAdminPopup',
-  },
-});
-
-// Order is important for the legend
-// WARNING: use UPPER CASE for the key
-const shortToLongNameOther = {
-  AB: 'Appenzeller Bahnen',
-  asm: 'Aare Seeland mobil AG',
-  AVA: 'Aargau Verkehr AG',
-  DB: 'Deutsche Bahn AG',
-  ÖBB: 'Österreichische Bundesbahnen',
-  RB: 'Rigi Bahnen',
-  RBS: 'Regionalverkehr Bern-Solothurn AG',
-  RhB: 'Rhätische Bahn AG',
-  SEHR: 'Stein am Rhein-Etzwilen-Hemishofen-Ramsen-Bahn',
-  SZU: 'Sihltal Zürich Uetliberg Bahn SZU AG',
-  TL: 'Transports publics de la région lausannoise',
-  VVT: 'Vapeur Val-de-Travers',
-  zb: 'Zentralbahn AG',
-};
-
-export const isbOther = new MapboxStyleLayer({
-  name: 'ch.sbb.isb.other',
-  mapboxLayer: netzkarteIsb,
-  styleLayersFilter: ({ metadata }) =>
-    metadata &&
-    metadata['isb.filter'] &&
-    /^(other|other_flag)$/.test(metadata['isb.filter']),
-  queryRenderedLayersFilter: ({ metadata }) =>
-    metadata &&
-    metadata['isb.filter'] &&
-    /^other$/.test(metadata['isb.filter']),
-  properties: {
-    hasInfos: true,
-    useOverlay: true,
-    popupComponent: 'IsbPopup',
-    layerInfoComponent: 'IsbOtherLayerInfo',
-    shortToLongName: shortToLongNameOther,
-    defaultColor: 'rgba(153, 71, 241, 1)', // Must be the same as in the style.
-  },
-});
-
-// Order is important for the legend
-// WARNING: use UPPER CASE for the key
-const shortToLongNameTVS = {
-  'SBB CFF FFS': 'SBB Infrastruktur',
-  BLS: 'BLS Netz AG',
-  SOB: 'SOB AG Infrastruktur',
-  CJ: 'Compagnie des chemins de fer du Jura SA',
-  ETB: 'Emmentalbahn GmbH',
-  HBS: 'Hafenbahn Schweiz AG',
-  OeBB: 'Oensingen-Balsthal-Bahn',
-  ST: 'Sursee-Triengen-Bahn',
-  STB: 'Sensetalbahn AG',
-  SZU: 'Sihltal Zürich Uetliberg Bahn SZU AG',
-  TMR: 'Transports de Martigny et Régions SA',
-  TPF: 'Transports publics fribourgeois SA',
-  TRAVYS: 'Travys',
-  TRN: 'Transports Publics Neuchâtelois',
-};
-
-export const isbTVS = new MapboxStyleLayer({
-  name: 'ch.sbb.isb.tvs',
-  mapboxLayer: netzkarteIsb,
-  styleLayersFilter: ({ metadata }) =>
-    metadata &&
-    metadata['isb.filter'] &&
-    /^(tvs|tvs_flag)$/.test(metadata['isb.filter']),
-  queryRenderedLayersFilter: ({ metadata }) =>
-    metadata && metadata['isb.filter'] && /^tvs$/.test(metadata['isb.filter']),
-  properties: {
-    hasInfos: true,
-    useOverlay: true,
-    popupComponent: 'IsbPopup',
-    layerInfoComponent: 'IsbTVSLayerInfo',
-    shortToLongName: shortToLongNameTVS,
-    defaultColor: 'rgba(0,91,169 , 1)', // Must be the same as in the style.
-    colors: {
-      'SBB CFF FFS': 'rgba(209, 1, 7, 1)',
-      BLS: 'rgba(53,164,48, 1)',
-      SOB: 'rgba(195, 156, 54, 1)',
-    },
-  },
-});
-
-export const geschosseLayer = new Layer({
-  name: 'ch.sbb.geschosse',
-  visible: true,
-});
-
-geschosseLayer.children = [-4, -3, -2, -1, 0, '2D', 1, 2, 3, 4].map((level) => {
-  return new LevelLayer({
-    name: `ch.sbb.geschosse${level}`,
-    visible: level === '2D',
-    mapboxLayer: dataLayer,
-    isQueryable: false,
-    styleLayersFilter: ({ metadata }) => metadata && metadata['geops.filter'],
-    level,
-    properties: {
-      radioGroup: 'ch.sbb.geschosse-layer',
-      parent: geschosseLayer,
-    },
-  });
 });
