+++ src/layerInfos/IsbOtherLayerInfo/IsbOtherLayerInfo.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
-import { isbOther } from '../../config/layers';
+import { isbOther } from '../../config/ch.sbb.isb';
 import IsbOtherLayerInfo from '.';
 
 describe('IsbOtherLayerInfo', () => {
